backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.10.0
django-extensions==3.2.3
PyJWT==2.8.0
pytz==2023.3
python-dateutil==2.9.0
sqlparse==0.4.4
typing-extensions==4.7.1